+++ src/index.js
@@ -2,21 +2,41 @@ import style from "./main.scss";
 import UIkit from 'uikit';
 import Icons from 'uikit/dist/js/uikit-icons';
 
-import { login, logout, create_account, send_verification } from './components/firebase/authentication/authentication';
-import { logIn as socialMediaLogIn, logOut as socialMediaLogOut, checkAuth, notifyAuthStateChanged } from './components/firebase/authentication/social_media/common_auth';
+import firebase from 'firebase';
+import firebaseConfig from './components/firebase/firebase_config.js';
+
+import { logIn, createAccount, sendVerification } from './components/firebase/authentication/traditional/auth.js';
+import socialMediaLogIn from './components/firebase/authentication/social_media/auth.js';
+import { logOut, checkAuth, notifyAuthStateChanged} from './components/firebase/authentication/common_auth.js';
 
 // components can be called from the imported UIkit reference
 // loads the Icon plugin
 UIkit.use(Icons);
 
+// Initialize Firebase
+firebase.initializeApp(firebaseConfig);
+
+
 document.querySelector('#googleLogIn').addEventListener('click', () => socialMediaLogIn('google'));
 document.querySelector('#githubLogIn').addEventListener('click', () => socialMediaLogIn('github'));
 document.querySelector('#facebookLogIn').addEventListener('click', () => socialMediaLogIn('facebook'));
 
-document.querySelector('#logOut').addEventListener('click', socialMediaLogOut);
+document.querySelector('#logOut').addEventListener('click', logOut);
 document.querySelector('#checkAuth').addEventListener('click', () => {
     UIkit.notification(checkAuth() ? 'Logged in' : 'Signed off');
 });
+
+/*
+    TODO: implement
+    logIn(userEmail, userPass)
+    createAccount(userEmail, userPass)
+    sendVerification()
+*/
+
+// document.querySelector('#login_field').addEventListener('click', logIn);
+// document.querySelector('#logout_field').addEventListener('click', logOut);
+// document.querySelector('#signUp').addEventListener('click', createAccount);
+// document.querySelector('#verification_field').addEventListener('click', sendVerification);
 
 notifyAuthStateChanged(function (user) {
     if (user) {
@@ -25,9 +45,4 @@ notifyAuthStateChanged(function (user) {
     } else {
         UIkit.notification('Signed off.');
     }
-})
-
-document.querySelector('#login_field').addEventListener('click', login);
-document.querySelector('#logout_field').addEventListener('click', logout);
-document.querySelector('#signUp').addEventListener('click', create_account);
-document.querySelector('#verification_field').addEventListener('click', send_verification);
+});
